added budget and flagging functionality and logo and new money bag animation
