+++ client/src/types/budget.ts
@@ -1,11 +1,6 @@
-export interface Budget {
-    food: number;
-    gas: number;
-    rent: number;
-    clothes: number;
-    recreation: number;
-    utilities: number;
-    otherAmount: number;
-    otherNote: string;
-  }
+// types/budget.ts
+
+export type Budget = {
+    total: number;
+  };
