+++ client/src/types/budget1.ts
@@ -1,0 +1,11 @@
+export interface Budget {
+    food: number;
+    gas: number;
+    rent: number;
+    clothes: number;
+    recreation: number;
+    utilities: number;
+    otherAmount: number;
+    otherNote: string;
+  }
+
